+++ src/codec/request/execute.rs
@@ -3,7 +3,7 @@ use std::io::Write;
 use crate::{
     codec::{
         consts::{keys, RequestType},
-        utils::{write_kv_str, write_kv_tuple, write_kv_u32},
+        utils::{write_kv_str, write_kv_tuple, write_kv_u64},
     },
     errors::EncodingError,
     tuple::Tuple,
@@ -13,7 +13,7 @@ use super::Request;
 
 #[derive(Clone, Debug)]
 enum ExecuteStatment<'a> {
-    StatementId(u32),
+    StatementId(u64),
     Query(&'a str),
 }
 
@@ -35,7 +35,7 @@ impl<'a, T: Tuple> Request for Execute<'a, T> {
     fn encode(&self, mut buf: &mut dyn Write) -> Result<(), EncodingError> {
         rmp::encode::write_map_len(&mut buf, 2)?;
         match self.statement {
-            ExecuteStatment::StatementId(x) => write_kv_u32(buf, keys::SQL_STMT_ID, x)?,
+            ExecuteStatment::StatementId(x) => write_kv_u64(buf, keys::SQL_STMT_ID, x)?,
             ExecuteStatment::Query(x) => write_kv_str(buf, keys::SQL_TEXT, x)?,
         }
         write_kv_tuple(buf, keys::SQL_BIND, &self.binds)?;
@@ -44,7 +44,7 @@ impl<'a, T: Tuple> Request for Execute<'a, T> {
 }
 
 impl<'a, T> Execute<'a, T> {
-    pub(crate) fn new_statement_id(statment_id: u32, binds: T) -> Self {
+    pub(crate) fn new_statement_id(statment_id: u64, binds: T) -> Self {
         Self {
             statement: ExecuteStatment::StatementId(statment_id),
             binds,