+++ src/codec/request/mod.rs
@@ -5,8 +5,8 @@ use crate::errors::EncodingError;
 
 pub(crate) use self::{
     auth::Auth, begin::Begin, call::Call, commit::Commit, delete::Delete, eval::Eval,
-    execute::Execute, id::Id, insert::Insert, ping::Ping, replace::Replace, rollback::Rollback,
-    select::Select, update::Update, upsert::Upsert,
+    execute::Execute, id::Id, insert::Insert, ping::Ping, prepare::Prepare, replace::Replace,
+    rollback::Rollback, select::Select, update::Update, upsert::Upsert,
 };
 
 use std::io::Write;
@@ -23,6 +23,7 @@ mod execute;
 mod id;
 mod insert;
 mod ping;
+mod prepare;
 mod replace;
 mod rollback;
 mod select;